Crate bftgrid_core

An actor framework geared towards BFT ordering libraries with support for deterministic simulation testing. It also allows writing a generic overall system construction logic independent from the specific actor system used. Every actor is managed by a single actor system but can interact with actors managed by other actor systems.

Structs

MessageNotSupported

Enums

ActorControl

P2PNetworkError

Traits

ActorMsg

ActorRef

An ActorRef can asynchronously send messages to the underlying actor, optionally with a delay, and create new actor references. Actor implementations can use ActorRefs to send messages to themselves and to other actors.

ActorSystem

An ActorSystem allows spawning actors by creating an [ActoRef] and setting its handler. The handler of the underlying actor can also be changed at any time. An ActorSystem can be cloned to obtain new references to it. Actors can use ActorSystem references to spawn new actors and even to change their own handlers.

Joinable

A Joinable can be awaited for completion. Specific types of ActorRef and ActorSystem can be joined to wait for their completion.

P2PNetwork

A P2PNetwork allows sending messages to other nodes in a P2P network.

TypedHandler

A TypedHandler is an actor that can handle messages of a specific type and optionally return an ActorControl message.

UntypedHandler

An UntypedHandler is an actor handler that can receive messages of any type, although it may refuse to handle some of them.

Type Aliases

AnActorMsg

AnActorRef

P2PNetworkResult

UntypedHandlerBox