bftgrid_core

Trait ActorSystem

pub trait ActorSystem: Clone {
    type ActorRefT<MsgT, HandlerT>: ActorRef<MsgT, HandlerT>
       where MsgT: ActorMsg,
             HandlerT: TypedHandler<MsgT = MsgT> + 'static;

    // Required methods
    fn create<MsgT, HandlerT>(
        &mut self,
        node_id: impl Into<String>,
        name: impl Into<String>,
    ) -> Self::ActorRefT<MsgT, HandlerT>
       where MsgT: ActorMsg,
             HandlerT: TypedHandler<MsgT = MsgT>;
    fn set_handler<MsgT, HandlerT>(
        &mut self,
        actor_ref: &mut Self::ActorRefT<MsgT, HandlerT>,
        handler: HandlerT,
    )
       where MsgT: ActorMsg,
             HandlerT: TypedHandler<MsgT = MsgT>;
}

An ActorSystem allows spawning actors by creating an [ActoRef] and setting its handler. The handler of the underlying actor can also be changed at any time. An ActorSystem can be cloned to obtain new references to it. Actors can use ActorSystem references to spawn new actors and even to change their own handlers.

Required Associated Types

type ActorRefT<MsgT, HandlerT>: ActorRef<MsgT, HandlerT> where MsgT: ActorMsg, HandlerT: TypedHandler<MsgT = MsgT> + 'static

Required Methods

fn create<MsgT, HandlerT>( &mut self, node_id: impl Into<String>, name: impl Into<String>, ) -> Self::ActorRefT<MsgT, HandlerT>
where MsgT: ActorMsg, HandlerT: TypedHandler<MsgT = MsgT>,

fn set_handler<MsgT, HandlerT>( &mut self, actor_ref: &mut Self::ActorRefT<MsgT, HandlerT>, handler: HandlerT, )
where MsgT: ActorMsg, HandlerT: TypedHandler<MsgT = MsgT>,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.